Two images uploading feature added

+++ application/controllers/Products.php
@@ -98,132 +98,271 @@ class Products extends CI_Controller {
     }
 
     public function add(){
-       
-        $data['title'] = "Rahul Textiles - Products";
+        if($this->session->has_userdata('adminusername')) {
+            $data['title'] = "Rahul Textiles - Products";
 
-        $this->form_validation->set_rules('name','Name',array('required'));
-        $this->form_validation->set_rules('description','Description',array('required'));
-        $this->form_validation->set_rules('price','Price',array('required'));
+            $this->form_validation->set_rules('name', 'Name', array('required'));
+            $this->form_validation->set_rules('description', 'Description', array('required'));
+            $this->form_validation->set_rules('price', 'Price', array('required'));
 
-        $name = $this->input->post('name');
-        $description = $this->input->post('description');
-        $price = $this->input->post('price');
-        $product_id = substr($name, 0, 3).'-'.date('His').'-'.date('dMY');
+            $name = $this->input->post('name');
+            $description = $this->input->post('description');
+            $price = $this->input->post('price');
+            $product_id = substr($name, 0, 3) . '-' . date('His') . '-' . date('dMY');
 
-        $this->load->view('header',$data);
-        $flag= "";
-        if ($this->form_validation->run() === FALSE)
-        {
-            $this->load->view('addproduct');
-        }
-        else
-        {
-                if($this->products_model->add_product($product_id,$name,$description,$price))
-                {
+            $this->load->view('header', $data);
+            $flag = "";
+            if ($this->form_validation->run() === FALSE) {
+                $this->load->view('addproduct');
+            } else {
+                if ($this->products_model->add_product($product_id, $name, $description, $price)) {
                     $flag = "success";
-                }
-                else
-                {
+                } else {
                     $flag = "Failed to add data in database";
                 }
+            }
+            if ($flag == "success") {
+                $uri = "/index.php/products/addimage/" . $product_id;
+                redirect($uri);
+            } else {
+                $this->load->view('addproduct', $flag);
+            }
+            $this->load->view('footer');
         }
-        if($flag == "success")
-        {
-//            $data['product_id'] = $product_id;
-//            $this->load->view('productimage',$data);
-            $uri = "/index.php/products/addimage/".$product_id;
-            redirect($uri);
-        } 
-        else
-        {
-            $this->load->view('addproduct',$flag);
+        else {
+            redirect(base_url('/index.php/admin'));
         }
-        $this->load->view('footer');
 	}
 
 	public function addimage($product_id = NULL){
-        $data['title'] = "Rahul Textiles - Products";
+        if($this->session->has_userdata('adminusername')) {
+            $data['title'] = "Rahul Textiles - Products";
 
-            if($this->input->post('product_id')!=NULL) {
+            if ($this->input->post('product_id') != NULL) {
+
                 $product_id = $this->input->post('product_id');
-                $config['upload_path'] = './uploads/';
-                $config['allowed_types'] = 'jpeg|jpg|png';
-                $config['max_size'] = 100;
-                $config['max_width'] = 1024;
-                $config['max_height'] = 768;
+                $imageCount = count(array_filter($_FILES['userfile']['name']));
 
-                $this->load->library('upload', $config);
-                $this->form_validation->set_rules('userfile', 'Image', array('required'));
-
-
-                if (!$this->upload->do_upload('userfile')) {
-                    $data['error'] = $this->upload->display_errors();
+                if ($imageCount < 1) {
+                    $data['error'] = "The Image field is required";
                 } else {
-                    $imageData = $this->upload->data();
-                    $imageName = $imageData['orig_name'];
-                    if ($this->products_model->update_product_image($imageName, $product_id)) {
-                        redirect(base_url('index.php/admin/dashboard'));
+                    if ($imageCount > 2) {
+                        $data['error'] = "Maximum 2 images can upload";
                     } else {
-                        $data['error'] = 'Failed to update database';
+                        $imageNames = array();
+                        $imageNames[0] = "";
+                        $imageNames[1] = "";
+                        for ($i = 0; $i < $imageCount; $i++) {
+                            $config['upload_path'] = './uploads/';
+                            $config['allowed_types'] = 'jpeg|jpg|png';
+//                            $config['max_size'] = 1000;
+//                            $config['max_width'] = 1024;
+//                            $config['max_height'] = 768;
+
+
+                            $_FILES['file']['name'] = $_FILES['userfile']['name'][$i];
+                            $_FILES['file']['type'] = $_FILES['userfile']['type'][$i];
+                            $_FILES['file']['tmp_name'] = $_FILES['userfile']['tmp_name'][$i];
+                            $_FILES['file']['error'] = $_FILES['userfile']['error'][$i];
+                            $_FILES['file']['size'] = $_FILES['userfile']['size'][$i];
+
+                            $imageName = time() . $_FILES['file']['name'];
+                            $config['file_name'] = $imageName;
+
+                            $this->load->library('upload', $config);
+                            $this->upload->initialize($config);
+
+                            if (!$this->upload->do_upload('file')) {
+                                $data['error'] = $this->upload->display_errors();
+                            } else {
+                                $imageData = $this->upload->data();
+                                $imageNames[$i] = $imageName;
+                            }
+                        }
+                        if ($this->products_model->update_product_image($imageNames[0], $imageNames[1], $product_id)) {
+                            redirect(base_url('index.php/products/addimage/' . $product_id));
+                        } else {
+                            var_dump($imageNames[0] . " - " . $imageNames[1]);
+                            $data['error'] = 'Failed to update database';
+                        }
                     }
                 }
             }
 
 
-        $this->load->view('header',$data);
-	    if($product_id != NULL){
-            $data['product'] = $this->products_model->get_product($product_id);
-            $this->load->view('productimage', $data);
+            $this->load->view('header', $data);
+            if ($product_id != NULL) {
+                $data['product'] = $this->products_model->get_product($product_id);
+                $this->load->view('productimage', $data);
+            } else {
+                $data['heading'] = "Product not selected";
+                $data['description'] = "Data is not retrived from database";
+                $this->load->view('empty_view', $data);
+            }
+            $this->load->view('footer');
         }
-	    else {
-            $data['heading'] = "Product not selected";
-            $data['description'] = "Data is not retrived from database";
-            $this->load->view('empty_view',$data);
+        else {
+            redirect(base_url('/index.php/admin'));
         }
-        $this->load->view('footer');
     }
 
     public function updateimage($product_id = NULL){
-        $data['title'] = "Rahul Textiles - Products";
+        if($this->session->has_userdata('adminusername')) {
+            $data['title'] = "Rahul Textiles - Products";
 
-        if($this->input->post('product_id')!=NULL) {
-            $product_id = $this->input->post('product_id');
-            $config['upload_path'] = './uploads/';
-            $config['allowed_types'] = 'jpeg|jpg|png';
-            $config['max_size'] = 100;
-            $config['max_width'] = 1024;
-            $config['max_height'] = 768;
+            if ($this->input->post('product_id') != NULL) {
 
-            $this->load->library('upload', $config);
-            $this->form_validation->set_rules('userfile', 'Image', array('required'));
+                $product_id = $this->input->post('product_id');
+                $imageCount = count(array_filter($_FILES['userfile']['name']));
 
-
-            if (!$this->upload->do_upload('userfile')) {
-                $data['error'] = $this->upload->display_errors();
-            } else {
-                $imageData = $this->upload->data();
-                $imageName = $imageData['orig_name'];
-                if ($this->products_model->update_product_image($imageName, $product_id)) {
-                    $this->session->set_flashdata('product_image_message','Updation success');
-                    redirect(base_url('index.php/products/updateproduct/'.$product_id));
+                if ($imageCount < 1) {
+                    $data['error'] = "The Image field is required";
                 } else {
-                    $data['error'] = 'Failed to update database';
+                    if ($imageCount > 2) {
+                        $data['error'] = "Maximum 2 images can upload";
+                    } else {
+                        $imageNames = array();
+                        $imageNames[0] = "";
+                        $imageNames[1] = "";
+                        for ($i = 0; $i < $imageCount; $i++) {
+                            $config['upload_path'] = './uploads/';
+                            $config['allowed_types'] = 'jpeg|jpg|png';
+//                            $config['max_size'] = 1000;
+//                            $config['max_width'] = 1024;
+//                            $config['max_height'] = 768;
+
+
+                            $_FILES['file']['name'] = $_FILES['userfile']['name'][$i];
+                            $_FILES['file']['type'] = $_FILES['userfile']['type'][$i];
+                            $_FILES['file']['tmp_name'] = $_FILES['userfile']['tmp_name'][$i];
+                            $_FILES['file']['error'] = $_FILES['userfile']['error'][$i];
+                            $_FILES['file']['size'] = $_FILES['userfile']['size'][$i];
+
+                            $imageName = time() . $_FILES['file']['name'];
+                            $config['file_name'] = $imageName;
+
+                            $this->load->library('upload', $config);
+                            $this->upload->initialize($config);
+
+                            if (!$this->upload->do_upload('file')) {
+                                $data['error'] = $this->upload->display_errors();
+                            } else {
+                                $imageData = $this->upload->data();
+                                $imageNames[$i] = $imageName;
+                            }
+                        }
+                        if ($this->products_model->update_product_image($imageNames[0], $imageNames[1], $product_id)) {
+                            $this->session->set_flashdata('product_image_message', 'Images updates successfully');
+                            redirect(base_url('index.php/products/updateimage/' . $product_id));
+                        } else {
+                            var_dump($imageNames[0] . " - " . $imageNames[1]);
+                            $data['error'] = 'Failed to update database';
+                        }
+                    }
                 }
             }
-        }
 
 
-        $this->load->view('header',$data);
-        if($product_id != NULL){
-            $data['product'] = $this->products_model->get_product($product_id);
-            $this->load->view('updateproduct', $data);
+            $this->load->view('header', $data);
+            if ($product_id != NULL) {
+                $data['product'] = $this->products_model->get_product($product_id);
+                $this->load->view('updateproduct', $data);
+            } else {
+                $data['heading'] = "Product not selected";
+                $data['description'] = "Data is not retrived from database";
+                $this->load->view('empty_view', $data);
+            }
+            $this->load->view('footer');
         }
         else {
-            $data['heading'] = "Product not selected";
-            $data['description'] = "Data is not retrived from database";
-            $this->load->view('empty_view',$data);
+            redirect(base_url('/index.php/admin'));
         }
-        $this->load->view('footer');
+    }
+
+    public function updatefirstimage()
+    {
+        if($this->session->has_userdata('adminusername')) {
+            if ($this->input->post('product_id') != NULL) {
+
+                $product_id = $this->input->post('product_id');
+
+
+                $config['upload_path'] = './uploads/';
+                $config['allowed_types'] = 'jpeg|jpg|png';
+
+
+                $_FILES['file']['name'] = $_FILES['userfile']['name'];
+                $_FILES['file']['type'] = $_FILES['userfile']['type'];
+                $_FILES['file']['tmp_name'] = $_FILES['userfile']['tmp_name'];
+                $_FILES['file']['error'] = $_FILES['userfile']['error'];
+                $_FILES['file']['size'] = $_FILES['userfile']['size'];
+
+                $imageName = time() . $_FILES['file']['name'];
+                $config['file_name'] = $imageName;
+
+                $this->load->library('upload', $config);
+                $this->upload->initialize($config);
+
+                if (!$this->upload->do_upload('file')) {
+                    $data['error'] = $this->upload->display_errors();
+                } else {
+                    $imageData = $this->upload->data();
+                    $imageName = $imageData['orig_name'];
+                }
+            }
+            if ($this->products_model->update_first_image($product_id, $imageName)) {
+                $this->session->set_flashdata('product_image_message', 'Image updated successfully');
+                redirect(base_url('index.php/products/updateimage/' . $product_id));
+            } else {
+                $data['error'] = 'Failed to update database';
+            }
+        }
+        else {
+            redirect(base_url('/index.php/admin'));
+        }
+    }
+
+    public function updatesecondimage()
+    {
+        if($this->session->has_userdata('adminusername')) {
+            if ($this->input->post('product_id') != NULL) {
+
+                $product_id = $this->input->post('product_id');
+
+
+                $config['upload_path'] = './uploads/';
+                $config['allowed_types'] = 'jpeg|jpg|png';
+
+
+                $_FILES['file']['name'] = $_FILES['userfile']['name'];
+                $_FILES['file']['type'] = $_FILES['userfile']['type'];
+                $_FILES['file']['tmp_name'] = $_FILES['userfile']['tmp_name'];
+                $_FILES['file']['error'] = $_FILES['userfile']['error'];
+                $_FILES['file']['size'] = $_FILES['userfile']['size'];
+
+                $imageName = time() . $_FILES['file']['name'];
+                $config['file_name'] = $imageName;
+
+                $this->load->library('upload', $config);
+                $this->upload->initialize($config);
+
+                if (!$this->upload->do_upload('file')) {
+                    $data['error'] = $this->upload->display_errors();
+                } else {
+                    $imageData = $this->upload->data();
+                    $imageName = $imageData['orig_name'];
+                }
+            }
+            if ($this->products_model->update_second_image($product_id, $imageName)) {
+                $this->session->set_flashdata('product_image_message', 'Image updated successfully');
+                redirect(base_url('index.php/products/updateimage/' . $product_id));
+            } else {
+                $data['error'] = 'Failed to update database';
+            }
+        }
+        else {
+            redirect(base_url('/index.php/admin'));
+        }
     }
 
     public function updateproduct($product_id = NULL){
@@ -257,19 +396,22 @@ class Products extends CI_Controller {
     }
 
     public function deleteproduct($product_id = NULL) {
-	    if($product_id === NULL){
-	        $this->session->set_flashdata('delete_message','Invalid Product details');
-            redirect(base_url('index.php/admin/dashboard'));
+        if($this->session->has_userdata('adminusername')) {
+            if ($product_id === NULL) {
+                $this->session->set_flashdata('delete_message', 'Invalid Product details');
+                redirect(base_url('index.php/admin/dashboard'));
+            } else {
+                if ($this->products_model->delete_product($product_id)) {
+                    $this->session->set_flashdata('delete_message', 'Product deleted successfully');
+                    redirect(base_url('index.php/admin/dashboard'));
+                } else {
+                    $this->session->set_flashdata('delete_message', 'Product not deleted from database');
+                    redirect(base_url('index.php/admin/dashboard'));
+                }
+            }
         }
-	    else {
-	        if($this->products_model->delete_product($product_id)){
-                $this->session->set_flashdata('delete_message','Product deleted successfully');
-                redirect(base_url('index.php/admin/dashboard'));
-            }
-	        else {
-                $this->session->set_flashdata('delete_message','Product not deleted from database');
-                redirect(base_url('index.php/admin/dashboard'));
-            }
+        else {
+            redirect(base_url('/index.php/admin'));
         }
     }
 }
